requirements.txt
torch==2.4.0
torchvision==0.19.0
efficientnet_pytorch==0.7.1
opencv-python==4.10.0
Pillow==10.4.0
